+++ ateam_web_spring/src/main/java/frige/FrigeService.java
@@ -3,7 +3,7 @@ package frige;
 import java.util.List;
 
 public interface FrigeService {
-	List<FrigeVO> frige_list(int user_id);	//재료 리스트 정보 가져오기
+	List<FrigeVO> frige_list(long user_id);	//재료 리스트 정보 가져오기
 		//재료정보 수정하기
 		//재료정보 삭제
 }
